refactor suite namespace
