+++ BackendBench/scripts/get_tests_stat.py
@@ -11,9 +11,7 @@ This is a helper script to analyze the test suite and provide statistics about t
 import statistics
 
 import torch
-from BackendBench.facto_suite import FactoTestSuite
-from BackendBench.opinfo_suite import OpInfoTestSuite
-from BackendBench.torchbench_suite import TorchBenchTestSuite
+from BackendBench.suite import OpInfoTestSuite, TorchBenchTestSuite, FactoTestSuite
 from BackendBench.scripts.pytorch_operators import extract_operator_name
